asdas
asdasd

+++ app/src/main/java/com/example/oportunia/domain/repository/StudentRepository.kt
@@ -7,6 +7,6 @@ import com.example.oportunia.domain.model.Student
  */
 interface StudentRepository {
     suspend fun findAllStudents(): Result<List<Student>>
-    suspend fun findStudentById(studentId: Long): Result<Student>
-    suspend fun findStudentByIdUser(idUser: Long): Result<Student>
+    suspend fun findStudentById(studentId: Int): Result<Student>
+    suspend fun findStudentByIdUser(idUser: Int): Result<Student>
 }
